Reorganization changes (removed uneccessary folders)
