+++ package.json
@@ -4,7 +4,7 @@
   "description": "Environmental App for Cal Poly CSC308, winter 202, Professor Bruno da Silva",
   "main": "index.js",
   "scripts": {
-    "serve": "nodemon backend/server/app/index.js",
+    "serve": "nodemon backend/index.js",
     "test": "jest"
   },
   "author": "308Squad",
